requirements.txt
requests~=2.25.0
python-chess~=1.3.3